backend/requirements.txt
allure-python-commons==2.13.5
asgiref==3.8.1
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
decorator==5.1.1
Django==5.0.4
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.8
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-django==4.8.0
pytest-mock==3.14.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.32.3
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tornado==6.4
trio==0.25.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
zope.interface==6.3
